method_merge_stars_merging_radius_cells and method_accretion_kernel_radius_cells are now doubles rather than ints

+++ src/Enzo/enzo_EnzoMethodMergeStars.hpp
@@ -65,7 +65,7 @@ protected: // methods
   
 protected: // attributes
 
-  int merging_radius_cells_;
+  double merging_radius_cells_;
 };
 
 #endif /* EnzoMethodMergeStars */
